remove settings dep from ad_filter
