+++ src/admc/select_container_dialog.cpp
@@ -25,6 +25,7 @@
 #include "ad/ad_defines.h"
 #include "ad/ad_utils.h"
 #include "ad/ad_filter.h"
+#include "console.h"
 
 #include <QTreeView>
 #include <QVBoxLayout>
